pyproject.toml
[tool.poetry]
name = "membership-management"
version = "0.1.0"
description = ""
authors = ["Inkyeong <ik.kim@time-gate.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
fastapi = "^0.115.6"
pydantic = { extras = ["email"], version = "^2.10.4" }
sqlalchemy = "^2.0.36"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
mysql = "^0.0.3"
pymysql = "^1.1.1"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-jose = "^3.3.0"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
fastapi-permissions = "^0.2.7"
starlette-context = "^0.3.6"
cryptography = "^44.0.0"
structlog = "^25.1.0"
mypy = "^1.14.1"
aiohttp = "^3.11.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
